requirements.txt
# SolanaGuard Requirements

# Core dependencies
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
seaborn==0.12.2
networkx==3.1
requests==2.30.0
python-dotenv==1.0.0

# Data processing and analysis
scikit-learn==1.2.2
plotly==5.15.0

# Jupyter notebook
jupyter==1.0.0
notebook==6.5.4
ipywidgets==8.0.6

# Dashboard
streamlit==1.22.0

# Web3 specific libraries
base58==2.1.1
py-solana-api==0.1.3

# Visualization
pydot==1.4.2
graphviz==0.20.1

# Utilities
tqdm==4.65.0